requirements.txt
pyzbar==0.1.8
qrcode==6.1
Pillow==7.2.0
